Move HTTPGateway Helper class
- Move Connection helper class HTTPGateway to an individual file
- Implement feedback code changes from review
